move presentational components to the containers folder
